CSS: Hide the news feed rather than remove it
Before the JS loads, the news feed is kept in place but made invisible.
Any JS errors causing failure to remove the element, causes the news feed
to infinitely load because the event handlers are still in place to
check for scrolling. Since the news feed is always at 'the bottom', the
feed keeps checking for more content, endlessly.

Closes #20 (hopefully)